Improving of GUI
- Function of retrieveVersionDataForFile can work with dummy data
- edit showComment function which get comment from data(vector of
versionRec*) in MainWindow instead calling FileArchiver function
- modify the setAsReference function which will display information to
use if use click the initial file for this function. If not, it will
carry the steps such as get comment and so on
-retrieveVersion function which use RetrieveForm class to display the
dialog to get the path and file name wanted. Then combine these two
information before passing it to FileArchiver(retrieveVersion function)

- RetrieveForm class: modify the selectionDir function which just allow
user to select the directory path

- TableModel class: add one function “resetData(vector)” to allow reset
whole table in GUI

** haven’t refer to functions in other classes. I just try with dummy
data with the classes i simply created.